// src/app/models/employees.model.ts
export class Employee {
    id: number;
    name: string;
    gender: string;
    email?: string;
    phoneNumber?: number;
    contactPreference: string;
    dateOfBirth: Date;
    address: string;
    department: string;
    isActive: boolean;
    photoPath?: string;
    qualification: string;
    working: string;
    martialStatus: string;
}